+++ ldp/ldpsend.c
@@ -152,7 +152,7 @@ static void construct_packet(char *data, size_t sz)
   ip = ether_payload(data);
   ip_set_version(ip, 4);
   ip46_set_min_hdr_len(ip);
-  ip46_set_payload_len(ip, global_opts.payload_size + 8);
+  ip46_set_payload_len(ip, (uint16_t)(global_opts.payload_size + 8));
   ip46_set_dont_frag(ip, 1);
   ip46_set_id(ip, 0);
   ip46_set_ttl(ip, 64);
@@ -163,8 +163,8 @@ static void construct_packet(char *data, size_t sz)
   udp = ip46_payload(ip);
   udp_set_src_port(udp, global_opts.src_port);
   udp_set_dst_port(udp, global_opts.dst_port);
-  udp_set_total_len(udp, global_opts.payload_size + 8);
-  udp_set_cksum_calc(ip, 20, udp, global_opts.payload_size + 8);
+  udp_set_total_len(udp, (uint16_t)(global_opts.payload_size + 8));
+  udp_set_cksum_calc(ip, 20, udp, (uint16_t)(global_opts.payload_size + 8));
 }
 
 static void usage(const char *argv0)
@@ -259,7 +259,7 @@ int main(int argc, char **argv)
         {
           usage(argv[0]);
         }
-        global_opts.src_port = uli;
+        global_opts.src_port = (uint16_t)uli;
         break;
       case 'n':
         uli = strtoul(optarg, &endptr, 10);
@@ -267,7 +267,7 @@ int main(int argc, char **argv)
         {
           usage(argv[0]);
         }
-        global_opts.dst_port = uli;
+        global_opts.dst_port = (uint16_t)uli;
         break;
       case 'i':
         uli = strtoul(optarg, &endptr, 10);
@@ -275,7 +275,7 @@ int main(int argc, char **argv)
         {
           usage(argv[0]);
         }
-        global_opts.interval_usec = uli;
+        global_opts.interval_usec = (int)uli;
         break;
       case 'c':
         uli = strtoul(optarg, &endptr, 10);
@@ -299,7 +299,7 @@ int main(int argc, char **argv)
         {
           usage(argv[0]);
         }
-        global_opts.burst_size = uli;
+        global_opts.burst_size = (int)uli;
         break;
       case 'h':
         usage(argv[0]);
@@ -332,7 +332,7 @@ int main(int argc, char **argv)
   for (i = 0; i < global_opts.burst_size; i++)
   {
     pkt_tbl[i].data = pkt;
-    pkt_tbl[i].sz = global_opts.payload_size + 8 + 20 + 14;
+    pkt_tbl[i].sz = (uint32_t)(global_opts.payload_size + 8 + 20 + 14);
   }
 
   while (global_opts.num_pkt <= 0 || pkts <= (size_t)global_opts.num_pkt)
@@ -341,7 +341,7 @@ int main(int argc, char **argv)
     int left = global_opts.burst_size;
     if (global_opts.num_pkt && left > (int)(global_opts.num_pkt - pkts))
     {
-      left = global_opts.num_pkt - pkts;
+      left = (int)(global_opts.num_pkt - pkts);
     }
     if (left == 0)
     {
@@ -361,9 +361,9 @@ int main(int argc, char **argv)
     {
       uint64_t pdiff = pkts - last_pkts;
       uint64_t bdiff = bytes - last_bytes;
-      double tdiff = time64 - last_time64;
+      double tdiff = (double)(time64 - last_time64);
       printf("%g MPPS %g Gbps\n",
-             pdiff/tdiff, bdiff*8/1000/tdiff);
+             (double)pdiff/tdiff, (double)bdiff*8/1000/tdiff);
       last_time64 = time64;
       last_pkts = pkts;
       last_bytes = bytes;
